+++ Cart/Products/Product.cs
@@ -15,26 +15,30 @@ public record class Product : IComparable<Product>
     /// <summary>
     /// Id.
     /// </summary>
-    public ulong Id { get; set; } // Is guid better?
+    public uint Id { get; set; }
+
     /// <summary>
-    /// Íàçâàíèå.
+    /// Наименование.
     /// </summary>
     public string Name { get; set; }
+
     /// <summary>
     /// Вес.
     /// </summary>
     public double Weight { get; set; }
+
     /// <summary>
     /// Стоимость.
     /// </summary>
     public decimal Price { get; set; }
+
     /// <summary>
     /// Дата доставки.
     /// </summary>
     public DateTime TimeOfArrival { get; set; }
 
     [JsonConstructor]
-    protected Product(ulong id, string name, double weight, decimal price, DateTime timeOfArrival)
+    protected Product(uint id, string name, double weight, decimal price, DateTime timeOfArrival)
     {
         Id = id;
         Name = name;
