+++ Cart/Products/WashingMachine.cs
@@ -9,7 +9,7 @@ public record WashingMachine : Product
 {
     public bool IsDryerIncluded { get; set; }
 
-    public WashingMachine(ulong id, string name, double weight, decimal price, DateTime timeOfArrival, bool isDryerIncluded) : base(id, name, weight, price, timeOfArrival)
+    public WashingMachine(uint id, string name, double weight, decimal price, DateTime timeOfArrival, bool isDryerIncluded) : base(id, name, weight, price, timeOfArrival)
     {
         IsDryerIncluded = isDryerIncluded;
     }
